+++ database/factories/BookFactory.php
@@ -6,7 +6,7 @@ use Faker\Generator as Faker;
 
 $factory->define(App\Models\Book::class, function (Faker $faker) {
     return [
-        'name' => $faker->name,
+        'name' => $faker->word,
         'price' => $faker->randomFloat(),
         'product_code' => $faker->word,
     ];
