+++ database/factories/StudentFactory.php
@@ -4,17 +4,6 @@ use App\Models\Student;
 use App\Models\User;
 use Faker\Generator as Faker;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
-
 $factory->define(Student::class, function (Faker $faker) {
     return [
         'idnumber' => $faker->randomNumber($nbDigits = 8),
