+++ tests/Feature/Http/Controllers/CommentControllerTest.php
@@ -18,14 +18,14 @@ class CommentControllerTest extends TestCase
      */
     public function destroy_returns_an_ok_response()
     {
-$this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
+        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
 
-$comment = factory(\App\Models\Comment::class)->create();
+        $comment = factory(\App\Models\Comment::class)->create();
 
-$response = $this->delete('comment/{comment}');
+        $response = $this->delete('comment/{comment}');
 
-$response->assertOk();
-$this->assertDeleted($);
+        $response->assertOk();
+        $this->assertDeleted($comment);
 
         // TODO: perform additional assertions
     }
@@ -35,15 +35,15 @@ $this->assertDeleted($);
      */
     public function store_returns_an_ok_response()
     {
-$this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
+        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
 
 
 
-$response = $this->post(route('storeComment'), [
-            // TODO: send request data
-        ]);
+        $response = $this->post(route('storeComment'), [
+                    // TODO: send request data
+                ]);
 
-$response->assertOk();
+        $response->assertOk();
 
         // TODO: perform additional assertions
     }
