+++ tests/Feature/Http/Controllers/GradeControllerTest.php
@@ -43,7 +43,6 @@ $this->markTestIncomplete('This test case was generated by Shift. When you are r
 $response = $this->delete('grades');
 
 $response->assertOk();
-$this->assertDeleted($);
 
         // TODO: perform additional assertions
     }
